Refactor all vscode code in to one file
This is to maximise code reuse with the lean-web-editor project.
All code that is specific to vscode is now in extension.ts.
So it should only be necessary to implement this file to get the system
to work in lean-web-editor.

Additionally:

- index.tsx has been split in to Main.tsx and index.tsx. index.tsx is
  vscode specific
- Main has been split in to `Main` and `InfoView` components. This is to
  separate the stateful parts of Main.
- onPin is now optional, if onPin is not provided then Info will not
  show the pin button
- rename DOM id `react_root` to `infoview_root`
