+++ infoview/Main.tsx
@@ -1,0 +1,134 @@
+import { PositionEvent, ConfigEvent, SyncPinEvent, TogglePinEvent, currentConfig, globalCurrentLoc,
+    ToggleAllMessagesEvent, syncPin, ServerStatus, Location, Config, defaultConfig, currentAllMessages,
+    AllMessagesEvent, PinnedLocation, locationEq} from './extension';
+import * as React from 'react';
+import { Message } from 'lean-client-js-core';
+import { Info } from './info';
+import { Messages, processMessages, ProcessedMessage } from './messages';
+import { Details } from './collapsing';
+import { useEvent } from './util';
+import { ContinueIcon, PauseIcon } from './svg_icons';
+import './tachyons.css' // stylesheet assumed by Lean widgets. See https://tachyons.io/ for documentation
+import './index.css'
+
+export const ConfigContext = React.createContext<Config>(defaultConfig);
+export const LocationContext = React.createContext<Location | null>(null);
+
+function StatusView(props: ServerStatus) {
+    return <Details>
+        <summary className="mv2 pointer">Tasks</summary>
+        <p>Running: {props.isRunning}</p>
+        <table> <tbody>
+            <tr key="header"><th>File Name</th>
+                <th>Pos start</th>
+                <th>Pos end</th>
+                <th>Desc</th></tr>
+            {props.tasks.map(t => <tr key={`${t.file_name}:${t.pos_col}:${t.pos_line}:${t.desc}`}>
+                <td>{t.file_name}</td>
+                <td>{t.pos_line}:{t.pos_col}</td>
+                <td>{t.end_pos_line}:{t.end_pos_col}</td>
+                <td>{t.desc}</td>
+            </tr>)}
+        </tbody>
+        </table>
+    </Details>
+}
+
+interface InfoViewProps {
+    config?: Config;
+    messages: Message[];
+    curLoc?: Location;
+}
+
+export function InfoView(props: InfoViewProps) {
+    if (!props) { return null }
+    const {config, curLoc, messages} = props;
+    if (!curLoc) return <p>Click somewhere in the Lean file to enable the info view.</p>;
+    const allMessages = processMessages(messages.filter((m) => curLoc && m.file_name === curLoc.file_name));
+    return <div className="ma1">
+        <ConfigContext.Provider value={config}>
+            <Infos curLoc={curLoc}/>
+            <div className="mv2"><AllMessages allMessages={allMessages}/></div>
+        </ConfigContext.Provider>
+    </div>
+}
+
+export function Main(props: {}) {
+    const [config, setConfig] = React.useState(currentConfig);
+    useEvent(ConfigEvent, (cfg) => setConfig(cfg), []);
+
+    const [messages, setMessages] = React.useState<Message[]>(currentAllMessages);
+    useEvent(AllMessagesEvent, (msgs) => setMessages(msgs), []);
+
+    const [curLoc, setCurLoc] = React.useState<Location>(globalCurrentLoc);
+    useEvent(PositionEvent, (loc) => setCurLoc(loc), []);
+
+    return <InfoView {...{config, messages, curLoc}}/>;
+}
+
+interface InfosProps {
+    curLoc: Location;
+}
+
+function Infos(props: InfosProps): JSX.Element {
+    const {curLoc} = props;
+    useEvent(SyncPinEvent, (syncMsg) => setPinnedLocs(syncMsg.pins), []);
+    useEvent(TogglePinEvent, () => isPinned(curLoc) ? unpin()() : pin());
+    const [pinnedLocs, setPinnedLocs] = React.useState<PinnedLocation[]>([]);
+    const isPinned = (loc: Location) => pinnedLocs.some((l) => locationEq(l, loc));
+    const pinKey = React.useRef<number>(0);
+    const pin = () => {
+        if (isPinned(curLoc)) {return; }
+        pinKey.current += 1;
+        const pins = [...pinnedLocs, { ...curLoc, key: pinKey.current }];
+        setPinnedLocs(pins);
+        syncPin(pins);
+    }
+    const unpin = (key?: number) => () => {
+        if (key === undefined) {
+            const pinned = pinnedLocs.find(p => locationEq(p, curLoc));
+            if (pinned) {
+                key = pinned.key;
+            } else {
+                return;
+            }
+        }
+        const pins = pinnedLocs.filter((l) => l.key !== key);
+        setPinnedLocs(pins);
+        syncPin(pins);
+    }
+    return <>
+        <div>
+            {pinnedLocs.map((loc) =>
+                <Info key={loc.key} loc={loc} isPinned={true} isCursor={false} onPin={unpin(loc.key)}/>)}
+        </div>
+        <Info loc={curLoc} isPinned={false} isCursor={true} onPin={pin}/>
+    </>;
+}
+
+function usePaused<T>(isPaused: boolean, t: T): T {
+    const old = React.useRef<T>(t);
+    if (!isPaused) old.current = t;
+    return old.current;
+}
+
+function AllMessages({allMessages: allMessages0}: {allMessages: ProcessedMessage[]}): JSX.Element {
+    const config = React.useContext(ConfigContext);
+    const [isPaused, setPaused] = React.useState<boolean>(false);
+    const allMessages = usePaused(isPaused, allMessages0);
+    const setOpenRef = React.useRef<React.Dispatch<React.SetStateAction<boolean>>>();
+    useEvent(ToggleAllMessagesEvent, () => setOpenRef.current((t) => !t));
+    return <Details setOpenRef={setOpenRef} initiallyOpen={!config.infoViewAutoOpenShowGoal}>
+        <summary>
+            All Messages ({allMessages.length})
+            <span className="fr">
+                <a className="link pointer mh2 dim"
+                        onClick={e => { e.preventDefault(); setPaused(!isPaused)}}
+                        title={isPaused ? 'continue updating' : 'pause updating'}>
+                    {isPaused ? <ContinueIcon/> : <PauseIcon/>}
+                </a>
+            </span>
+        </summary>
+        <div className="ml1"> <Messages messages={allMessages}/> </div>
+    </Details>;
+}
